requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.16
django-celery-beat==2.7.0
django-mssql-backend==2.8.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
mssql-django==1.5
prompt-toolkit==3.0.48
PyJWT==2.10.0
pyodbc==5.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
